+++ s/shopify/parts/options.ts
@@ -3,6 +3,7 @@ import {Remote} from "../remote/remote.js"
 import {ImageFormat} from "../graphql/units/image.js"
 import {CheckoutLineItem} from "../graphql/checkout.js"
 import {ProductFilters} from "../graphql/types/product_filters.js"
+import {ProductSortKeys} from "../graphql/types/product_sort_keys.js"
 import {ProductQuerySpec} from "../product_queries/convert_product_query_spec_to_string.js"
 
 export namespace Options {
@@ -20,6 +21,8 @@ export namespace Options {
 		page_size?: number
 		query?: ProductQuerySpec
 		image_format?: ImageFormat
+		sort?: ProductSortKeys
+		reverse?: boolean
 	}
 
 	export type Collections = {